requirements.txt
fastapi
uvicorn
pydantic
PyYAML
psycopg2-binary
openpyxl

sqlalchemy
nanoid

itsdangerous
asyncpg
python-multipart
google_auth_oauthlib

tzdata
apscheduler
